Fix include-related errors from clang-tidy.
Clang-tidy requires all relevant headers to be included, even if the
program compiles without a problem. This won't make a practical
difference in behavior, it's just to suppress errors.
