+++ src/backup.h
@@ -21,6 +21,9 @@
 #ifndef BACKUP_H_INCLUDED
 #define BACKUP_H_INCLUDED
 
+#include "base_types.h"
+#include <vector>
+
 #define UNC_BACKUP_SUFFIX        ".unc-backup~"
 #define UNC_BACKUP_MD5_SUFFIX    ".unc-backup.md5~"
