+++ src/char_table.h
@@ -10,6 +10,9 @@
 #ifndef CHAR_TABLE_H_INCLUDED
 #define CHAR_TABLE_H_INCLUDED
 
+#include "base_types.h"
+#include <cstddef>
+
 /**
  * bit0-7 = paired char
  * bit8 = OK for keyword 1st char
